object mapping is not used in task
WSDL generation is disfunctional if a classmap is used.
this patch fixes it, but if an object array is declared in the classmap and
in a method definition then the wsdl is disfunctional.

for instance you define in your app.yml

``` yaml
  ck_web_service_plugin:
    # ...
    soap_options:
      # ...
      classmap:
        Customer: User
        CustomerArray: ckGenericArray
```

and in a module

``` php
/**
 * @return Customer[] list of accessible clients
 */
function executeMyAction(sfWebRequest)
{
}
```

it will turn into a double entry in wsdl.